// src/app/rest/model/event-term-public.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EventTermPublicDTO { 
    id?: number;
    startAt?: string;
    endAt?: string;
    capacity: number;
    deposit: number;
    price: number;
}