// src/app/rest/model/registration-pricing-request.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RegistrationPricingRequestDTO { 
    eventTermId: number;
    userEmail: string;
    numberOfPeople: number;
}